requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-tenants==3.4.1
djangorestframework==3.13.1
environs==9.4.0
marshmallow==3.14.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
